Split `EmailTemplateValidator` to its own file
- Move all Email Notification tests to `/tests/test_email/`
- Refactor existing tests in `/tests/...` and remove duplicate tests
