+++ email_notification/sender/email_notification_builder.py
@@ -1,107 +1,21 @@
 import datetime
-import os
-import re
 from decimal import Decimal
 from typing import Any
 from urllib.parse import urlencode, urljoin
 
 from django.conf import settings
-from django.core.exceptions import ValidationError
-from django.core.files.uploadedfile import InMemoryUploadedFile
 from django.utils.timezone import get_default_timezone
-from jinja2 import TemplateError
-from jinja2.sandbox import SandboxedEnvironment
 
-from email_notification.exceptions import (
-    EmailBuilderConfigError,
-    EmailNotificationBuilderError,
-    EmailTemplateValidationError,
-)
+from email_notification.exceptions import EmailBuilderConfigError, EmailNotificationBuilderError
 from email_notification.models import EmailTemplate
 from email_notification.sender.email_notification_context import EmailNotificationContext
-from email_notification.templatetags.email_template_filters import format_currency
+from email_notification.sender.email_notification_validator import EmailTemplateValidator
 from reservations.models import Reservation
 from tilavarauspalvelu.utils.commons import LANGUAGES
 
-FILTERS_MAP = {"currency": format_currency}
-
-
-def get_sandboxed_environment() -> SandboxedEnvironment:
-    env = SandboxedEnvironment()
-
-    for fil, func in FILTERS_MAP.items():
-        env.filters[fil] = func
-
-    return env
-
-
-class EmailTemplateValidator:
-    @staticmethod
-    def _validate_tags(string: str) -> None:
-        # Matches "{{ word }}" or "{{ word1 | word2 }}".
-        # Note, this doesn't match the preferred format with spaces around the pipe: "{{word1|word2}}"
-        bracket_lookup = re.compile(r"{{ *(\w+) \| *(\w+) *}}|{{ *(\w+) *}}")
-
-        tags_inside_brackets = re.findall(bracket_lookup, string)
-        variable_tags = []
-
-        for strings in tags_inside_brackets:
-            strings_list = [tag for tag in strings if tag and tag not in FILTERS_MAP]
-
-            variable_tags.append(strings_list[0])
-
-        for tag in variable_tags:
-            if tag not in settings.EMAIL_TEMPLATE_CONTEXT_VARIABLES:
-                raise EmailTemplateValidationError(f"Tag '{tag}' is not supported")
-
-    @staticmethod
-    def _validate_illegals(string: str) -> None:
-        # Matches "{% word %}".
-        expression_lookup = re.compile(r"{% *(\w+) *")
-
-        expressions = re.findall(expression_lookup, string)
-        for expression in expressions:
-            if expression not in settings.EMAIL_TEMPLATE_SUPPORTED_EXPRESSIONS:
-                raise EmailTemplateValidationError(f"Illegal tags found: tag was {expression}")
-
-    @staticmethod
-    def _validate_in_sandbox(string: str, context: dict, env: SandboxedEnvironment) -> None:
-        try:
-            env.from_string(string).render(context)
-        except TemplateError as e:
-            raise EmailTemplateValidationError(e)
-
-    def validate_string(self, string: str, context_dict: dict = (), env: SandboxedEnvironment = None):
-        env = env or get_sandboxed_environment()
-        self._validate_in_sandbox(string, context_dict, env)
-        self._validate_illegals(string)
-        self._validate_tags(string)
-
-    def validate_html_file(self, value: InMemoryUploadedFile, context_dict=()) -> None:
-        # File extension
-        file_extension = os.path.splitext(value.name)[1]
-        if file_extension.lower() != ".html":
-            raise ValidationError(f"Unsupported file extension {file_extension}. Only .html files are allowed")
-
-        # File size
-        if value.size <= 0 or value.size > settings.EMAIL_HTML_MAX_FILE_SIZE:
-            raise ValidationError(
-                f"Invalid HTML file size. Allowed file size: 1-{settings.EMAIL_HTML_MAX_FILE_SIZE} bytes"
-            )
-
-        # File content
-        try:
-            file = value.open()
-            content = file.read().decode("utf-8")
-            self.validate_string(content, context_dict)
-        except EmailTemplateValidationError as err:
-            raise ValidationError(err.message)
-        except Exception as err:
-            raise ValidationError(f"Unable to read the HTML file: {err!s}")
-
 
 class ReservationEmailNotificationBuilder:
-    validator = EmailTemplateValidator
+    validator: EmailTemplateValidator
     template: EmailTemplate
     context: EmailNotificationContext
     reservation: Reservation | None
@@ -122,7 +36,7 @@ class ReservationEmailNotificationBuilder:
         self.context = context or EmailNotificationContext.from_reservation(reservation)
         self._set_language(language or self.context.reservee_language)
         self._init_context_attr_map()
-        self.env = get_sandboxed_environment()
+        self.validator = EmailTemplateValidator()
         self.validate_template()
 
     def _get_reservee_name(self) -> str:
@@ -257,26 +171,23 @@ class ReservationEmailNotificationBuilder:
             self.context_attr_map[key] = value()
 
     def validate_template(self) -> None:
-        validator = self.validator()
-
         html_content = self._get_html_content(self.template)
         if html_content:
-            validator.validate_string(html_content, self.context_attr_map, env=self.env)
+            self.validator.validate_string(html_content, self.context_attr_map)
 
-        validator.validate_string(self.template.subject, self.context_attr_map, env=self.env)
-        validator.validate_string(self.template.content, self.context_attr_map, env=self.env)
+        self.validator.validate_string(self.template.subject, self.context_attr_map)
+        self.validator.validate_string(self.template.content, self.context_attr_map)
 
     def get_subject(self) -> str:
         subject = self._get_by_language(self.template, "subject")
-        rendered = self.env.from_string(subject).render(self.context_attr_map)
-        return rendered
+        return self.validator.render_string(string=subject, context=self.context_attr_map)
 
     def get_content(self) -> str:
         content = self._get_by_language(self.template, "content")
-        return self.env.from_string(content).render(self.context_attr_map)
+        return self.validator.render_string(string=content, context=self.context_attr_map)
 
     def get_html_content(self) -> str | None:
         content = self._get_html_content(self.template)
-        if content:
-            return self.env.from_string(content).render(self.context_attr_map)
-        return None
+        if not content:
+            return None
+        return self.validator.render_string(string=content, context=self.context_attr_map)
