+++ email_notification/sender/email_notification_validator.py
@@ -1,0 +1,95 @@
+import os
+import re
+from typing import Any
+
+from django.conf import settings
+from django.core.exceptions import ValidationError
+from django.core.files.uploadedfile import InMemoryUploadedFile
+from jinja2 import TemplateError
+from jinja2.sandbox import SandboxedEnvironment
+
+from email_notification.exceptions import EmailTemplateValidationError
+from email_notification.templatetags.email_template_filters import format_currency
+
+FILTERS_MAP = {"currency": format_currency}
+
+
+class EmailTemplateValidator:
+    _env: None | SandboxedEnvironment
+
+    def __init__(self) -> None:
+        self._env = None
+
+    @property
+    def env(self) -> SandboxedEnvironment:
+        if self._env:
+            return self._env
+
+        env = SandboxedEnvironment()
+        for fil, func in FILTERS_MAP.items():
+            env.filters[fil] = func
+
+        self._env = env
+        return env
+
+    @staticmethod
+    def _validate_tags(string: str) -> None:
+        # Matches "{{ word }}" or "{{ word1 | word2 }}".
+        # Note, this doesn't match the preferred format with spaces around the pipe: "{{word1|word2}}"
+        bracket_lookup = re.compile(r"{{ *(\w+) \| *(\w+) *}}|{{ *(\w+) *}}")
+
+        tags_inside_brackets = re.findall(bracket_lookup, string)
+        variable_tags = []
+
+        for strings in tags_inside_brackets:
+            strings_list = [tag for tag in strings if tag and tag not in FILTERS_MAP]
+
+            variable_tags.append(strings_list[0])
+
+        for tag in variable_tags:
+            if tag not in settings.EMAIL_TEMPLATE_CONTEXT_VARIABLES:
+                raise EmailTemplateValidationError(f"Tag '{tag}' is not supported")
+
+    @staticmethod
+    def _validate_illegals(string: str) -> None:
+        # Matches "{% word %}".
+        expression_lookup = re.compile(r"{% *(\w+) *")
+
+        expressions = re.findall(expression_lookup, string)
+        for expression in expressions:
+            if expression not in settings.EMAIL_TEMPLATE_SUPPORTED_EXPRESSIONS:
+                raise EmailTemplateValidationError(f"Illegal tags found: tag was '{expression}'")
+
+    def render_string(self, string: str, context: dict[str, Any]) -> str:
+        """Render the given string with the given context in a safe, sandboxed environment."""
+        try:
+            return self.env.from_string(string).render(context)
+        except TemplateError as e:
+            raise EmailTemplateValidationError(e)
+
+    def validate_string(self, string: str, context_dict: dict[str, Any]):
+        self._validate_illegals(string)
+        self._validate_tags(string)
+        self.render_string(string, context_dict)
+
+    def validate_html_file(self, value: InMemoryUploadedFile, context_dict=()) -> None:
+        # File extension
+        file_extension = os.path.splitext(value.name)[1]
+        if file_extension.lower() != ".html":
+            raise ValidationError(f"Unsupported file extension {file_extension}. Only .html files are allowed")
+
+        # File size
+        if value.size <= 0 or value.size > settings.EMAIL_HTML_MAX_FILE_SIZE:
+            raise ValidationError(
+                f"Invalid HTML file size. Allowed file size: 1-{settings.EMAIL_HTML_MAX_FILE_SIZE} bytes"
+            )
+
+        # File content
+        try:
+            file = value.open()
+            content = file.read().decode("utf-8")
+            self.validate_string(content, context_dict)
+        except EmailTemplateValidationError as err:
+            raise ValidationError(err.message)
+        except Exception as err:
+            raise ValidationError(f"Unable to read the HTML file: {err!s}")
